requirements.txt
flask
keras
numpy
pillow
